Finish CreateTaskActivity, and fetch all task according to the project in the task list, I haven't complete the datepicker yet

+++ ContactaniserApp/src/csse3005/contactaniser/datasource/ProjectDataSource.java
@@ -110,7 +110,7 @@ public class ProjectDataSource {
 
 	private Project cursorToProject(Cursor cursor) {
 		Project project = new Project();
-		project.setProjectid(cursor.getInt(0));
+		project.setProjectid(cursor.getLong(0));
 		project.setProjectName(cursor.getString(1));
 		project.setProjectDescription(cursor.getString(2));
 		Date sd = Date.valueOf(cursor.getString(3));
